[cpu] split opcode implementations out of the main control module.
